small typo fixes and organization of frontend
